requirements.txt
requests
Pillow
customtkinter
tkinterdnd2
